Move lazy_caching and async_provide_credentials_fn to config
