+++ aws/rust-runtime/aws-auth-providers/src/profile/repr.rs
@@ -13,8 +13,8 @@
 //! multiple actions into the same profile).
 
 use crate::profile::ProfileFileError;
-use aws_auth::Credentials;
 use aws_types::profile::{Profile, ProfileSet};
+use aws_types::Credentials;
 
 /// Chain of Profile Providers
 ///
