+++ aws/rust-runtime/aws-auth/src/lib.rs
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+//! AWS authentication middleware used to store and retrieve credentials from the property bag
+
 pub mod middleware;
-pub mod provider;
 
 use aws_types::credentials::SharedCredentialsProvider;
 use smithy_http::property_bag::PropertyBag;
