+++ aws/rust-runtime/aws-sig-auth/src/event_stream.rs
@@ -4,10 +4,10 @@
  */
 
 use crate::middleware::Signature;
-use aws_auth::Credentials;
 use aws_sigv4::event_stream::{sign_empty_message, sign_message};
 use aws_sigv4::SigningParams;
 use aws_types::region::SigningRegion;
+use aws_types::Credentials;
 use aws_types::SigningService;
 use smithy_eventstream::frame::{Message, SignMessage, SignMessageError};
 use smithy_http::property_bag::{PropertyBag, SharedPropertyBag};
@@ -91,9 +91,9 @@ impl SignMessage for SigV4Signer {
 mod tests {
     use crate::event_stream::SigV4Signer;
     use crate::middleware::Signature;
-    use aws_auth::Credentials;
     use aws_types::region::Region;
     use aws_types::region::SigningRegion;
+    use aws_types::Credentials;
     use aws_types::SigningService;
     use smithy_eventstream::frame::{HeaderValue, Message, SignMessage};
     use smithy_http::property_bag::PropertyBag;
